+++ lib/ice_cube/validations/day_of_month.rb
@@ -13,9 +13,7 @@ module IceCube
       self
     end
 
-    class Validation
-
-      include Validations::Lock
+    class Validation < Validations::FixedValue
 
       attr_reader :day
       alias :value :day
@@ -26,6 +24,10 @@ module IceCube
 
       def type
         :day
+      end
+
+      def dst_adjust?
+        true
       end
 
       def build_s(builder)