+++ lib/ice_cube/validations/hour_of_day.rb
@@ -14,9 +14,7 @@ module IceCube
       self
     end
 
-    class Validation
-
-      include Validations::Lock
+    class Validation < Validations::FixedValue
 
       attr_reader :hour
       alias :value :hour
@@ -27,6 +25,10 @@ module IceCube
 
       def type
         :hour
+      end
+
+      def dst_adjust?
+        true
       end
 
       def build_s(builder)