+++ spec/spec_helper.rb
@@ -7,8 +7,11 @@ end
 
 require File.dirname(__FILE__) + '/../lib/ice_cube'
 
+IceCube.compatibility = 12
+
 DAY = Time.utc(2010, 3, 1)
 WEDNESDAY = Time.utc(2010, 6, 23, 5, 0, 0)
+
 WORLD_TIME_ZONES = [
   'America/Anchorage',  # -1000 / -0900
   'Europe/London',      # +0000 / +0100
@@ -16,6 +19,9 @@ WORLD_TIME_ZONES = [
 ]
 
 RSpec.configure do |config|
+  Dir[File.dirname(__FILE__) + '/support/**/*'].each { |f| require f }
+
+  config.include WarningHelpers
 
   config.around :each, :if_active_support_time => true do |example|
     example.run if defined? ActiveSupport
@@ -48,4 +54,11 @@ RSpec.configure do |config|
       end
     end
   end
+
+  config.around :each, expect_warnings: true do |example|
+    capture_warnings do
+      example.run
+    end
+  end
+
 end