+++ CHANGELOG.md
@@ -1,5 +1,32 @@
 # Changelog
 
+## 0.12.1 / 2014-07-04
+
+* [FEATURE]    Deprecation message improvements
+* [FEATURE]    Added support for deserialization of times via Time.parse
+* [FEATURE]    Added interval validations
+* [FEATURE]    Coerce validation intervals to Fixnum
+* [BUGFIX]     Fix YAML serialization on blank values in ActiveRecord (#231)
+* [BUGFIX]     Yearly interval should return self like others
+
+## 0.12.0 / 2014-04-06
+
+* [FEATURE]     Rename to `start_time` as a hash key (see UPGRADING) (#102)
+* [FEATURE]     Notify of deprecated usage (#219)
+* [BUGFIX]      Skip double occurrences over DST (#189)
+* [BUGFIX]      Avoid symbolizing hash keys from input
+* [BUGFIX]      Ensure time comparisons are done in schedule time zone (#209)
+* [BUGFIX]      Occurrence#overnight? now works on the last day of the month (#218)
+
+## 0.11.3 / 2014-02-07
+
+* [BUGFIX]      Fix a StopIteration leak
+
+## 0.11.2 / 2014-01-25
+
+* [ENHANCEMENT] Use Enumerator for schedule occurrences
+* [BUGFIX]      Fix high CPU usage on minutely schedules
+
 ## 0.11.1 / 2013-10-28
 
 * [ENHANCEMENT] Move deprecated into IceCube namespace